Update README to compile non-shared Fortran obj...
Update README to compile non-shared Fortran object.

Configure dynamically Cython extension to dynamically link to
`libcilkrts` to provide support for Cilk Plus reduction sum.  Note that
this requires `libcilkrts.so` to be available on the library path.  This
library is included in both the Intel compiler suite or the Cilk branch
of gcc.